Split read.rs into diferent files
