+++ examples/extract.rs
@@ -10,8 +10,7 @@ fn main() {
     std::process::exit(real_main());
 }
 
-fn real_main() -> i32
-{
+fn real_main() -> i32 {
     let args: Vec<_> = std::env::args().collect();
     if args.len() < 2 {
         println!("Usage: {} <filename>", args[0]);
@@ -22,15 +21,16 @@ fn real_main() -> i32
 
     let mut archive = zip::ZipArchive::new(file).unwrap();
 
-    for i in 0..archive.len()
-    {
+    for i in 0..archive.len() {
         let mut file = archive.by_index(i).unwrap();
         let outpath = sanitize_filename(file.name());
         println!("{}", outpath.display());
 
         {
             let comment = file.comment();
-            if comment.len() > 0 { println!("  File comment: {}", comment); }
+            if comment.len() > 0 {
+                println!("  File comment: {}", comment);
+            }
         }
 
         create_directory(outpath.parent().unwrap_or(std::path::Path::new("")), None);
@@ -40,8 +40,7 @@ fn real_main() -> i32
         if (&*file.name()).ends_with("/") {
             create_directory(&outpath, perms);
 
-        }
-        else {
+        } else {
             write_file(&mut file, &outpath, perms);
         }
     }
@@ -50,21 +49,18 @@ fn real_main() -> i32
 }
 
 #[cfg(unix)]
-fn convert_permissions(mode: Option<u32>) -> Option<fs::Permissions>
-{
+fn convert_permissions(mode: Option<u32>) -> Option<fs::Permissions> {
     match mode {
         Some(mode) => Some(fs::Permissions::from_mode(mode)),
         None => None,
     }
 }
 #[cfg(not(unix))]
-fn convert_permissions(_mode: Option<u32>) -> Option<fs::Permissions>
-{
+fn convert_permissions(_mode: Option<u32>) -> Option<fs::Permissions> {
     None
 }
 
-fn write_file(file: &mut zip::read::ZipFile, outpath: &std::path::Path, perms: Option<fs::Permissions>)
-{
+fn write_file(file: &mut zip::ZipFile, outpath: &std::path::Path, perms: Option<fs::Permissions>) {
     let mut outfile = fs::File::create(&outpath).unwrap();
     io::copy(file, &mut outfile).unwrap();
     if let Some(perms) = perms {
@@ -72,16 +68,14 @@ fn write_file(file: &mut zip::read::ZipFile, outpath: &std::path::Path, perms: O
     }
 }
 
-fn create_directory(outpath: &std::path::Path, perms: Option<fs::Permissions>)
-{
+fn create_directory(outpath: &std::path::Path, perms: Option<fs::Permissions>) {
     fs::create_dir_all(&outpath).unwrap();
     if let Some(perms) = perms {
         fs::set_permissions(outpath, perms).unwrap();
     }
 }
 
-fn sanitize_filename(filename: &str) -> std::path::PathBuf
-{
+fn sanitize_filename(filename: &str) -> std::path::PathBuf {
     let no_null_filename = match filename.find('\0') {
         Some(index) => &filename[0..index],
         None => filename,
@@ -91,7 +85,7 @@ fn sanitize_filename(filename: &str) -> std::path::PathBuf
         .components()
         .filter(|component| match *component {
             std::path::Component::Normal(..) => true,
-            _ => false
+            _ => false,
         })
         .fold(std::path::PathBuf::new(), |mut path, ref cur| {
             path.push(cur.as_os_str());
