+++ src/zip_file_data.rs
@@ -1,29 +1,5 @@
-//! Types that specify what is contained in a ZIP.
-
 use time;
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum System {
-    Dos = 0,
-    Unix = 3,
-    Unknown,
-    #[doc(hidden)]
-    __Nonexhaustive,
-}
-
-impl System {
-    pub fn from_u8(system: u8) -> System {
-        use self::System::*;
-
-        match system {
-            0 => Dos,
-            3 => Unix,
-            _ => Unknown,
-        }
-    }
-}
-
-pub const DEFAULT_VERSION: u8 = 20;
+use system::System;
 
 /// Structure representing a ZIP file.
 #[derive(Debug)]
@@ -54,16 +30,4 @@ pub struct ZipFileData {
     pub header_start: u64,
     /// External file attributes
     pub external_attributes: u32,
-}
-
-#[cfg(test)]
-mod test {
-    #[test]
-    fn system() {
-        use super::System;
-        assert_eq!(System::Dos as u16, 0u16);
-        assert_eq!(System::Unix as u16, 3u16);
-        assert_eq!(System::from_u8(0), System::Dos);
-        assert_eq!(System::from_u8(3), System::Unix);
-    }
 }
